Moved existing tests into modules (#127)
